FLX-689 Multiple funding sources in transactions preventing payment

+++ lib/extensions/models/fluxx_request_transaction.rb
@@ -331,9 +331,10 @@ module FluxxRequestTransaction
               rtfs.updated_by_id = self.updated_by_id
               rtfs.save
             end
-          elsif rtfs && rtfs.id
+          elsif rtfs && rtfs.id && !rtfs.new_record?
             # The user removed the value, let's delete the record as well
-            rtfs.destroy
+            rtfs = rtfs.reload rescue nil # Note that for state transitions we essentially save twice; once to update the values and second time to do the state transition
+            rtfs.destroy if rtfs
           end
         end
       end
